LayoutTests/fast/regions/content-webkit-from-flow-parsing-expected.txt

Tests being able to set content to -webkit-from-flow

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testCSSText("content: -webkit-from-flow('first-flow')") is "-webkit-from-flow(first-flow)"
PASS testCSSText("content: -webkit-from-flow('first flow')") is "-webkit-from-flow('first flow')"
PASS testCSSText("content: -webkit-from-flow('auto')") is "-webkit-from-flow(auto)"
PASS testCSSText("content: -webkit-from-flow(auto)") is ""
PASS testCSSText("content: -webkit-from-flow(initial)") is ""
PASS testCSSText("content: -webkit-from-flow(inherit)") is ""
PASS testCSSText("content: -webkit-from-flow()") is ""
PASS testCSSText("content: -webkit-from-flow('')") is "-webkit-from-flow('')"
PASS testCSSText("content: ;") is ""
PASS testCSSText("content: -webkit-from-flow(1)") is ""
PASS testCSSText("content: -webkit-from-flow(1.2)") is ""
PASS testCSSText("content: -webkit-from-flow(-1)") is ""
PASS testCSSText("content: -webkit-from-flow(12px)") is ""
PASS testCSSText("content: -webkit-from-flow(first flow)") is ""
PASS testCSSText("content: -webkit-from-flow(first-flow)") is ""
PASS testComputedStyle("-webkit-from-flow('first-flow')") is "first-flow"
PASS testComputedStyle("-webkit-from-flow('first flow')") is "'first flow'"
PASS testComputedStyle("-webkit-from-flow('auto')") is "auto"
PASS testComputedStyle("-webkit-from-flow(auto)") is ""
PASS testComputedStyle("-webkit-from-flow(initial)") is ""
PASS testComputedStyle("-webkit-from-flow(inherit)") is ""
PASS testComputedStyle("-webkit-from-flow()") is ""
PASS testComputedStyle("-webkit-from-flow(1)") is ""
PASS testComputedStyle("-webkit-from-flow(1.2)") is ""
PASS testComputedStyle("-webkit-from-flow(-1)") is ""
PASS testComputedStyle("-webkit-from-flow(12px)") is ""
PASS testComputedStyle("-webkit-from-flow('')") is "''"
PASS testComputedStyle("-webkit-from-flow(first flow)") is ""
PASS testComputedStyle("-webkit-from-flow(first-flow)") is ""
PASS successfullyParsed is true

TEST COMPLETE
